MAJOR: prevent the default homepage from redirecting to root
Extended RootURLController to be able to set is_at_root for locales
other then the homepage locale. Added a custom static
get_homepage_link_by_locale() function that returns the URLSegment
instead of the RelativeLink() because the latter is now empty for
homepages other then the default locale. It now also finds the first
page as homepage if it is not a translation of the default homepage.
The $_POST[] hack that prevented the redirect earlier has now become
obsolete.