FIX: make sure the correct locale is present in the EditForm
This is a temporary fix for Translatable not keeping time of the
correct locale for the page that is edited. After the page is first
saved, the locale disappears from the form url, so the CMS reverts to
the default language.

This locale is necessary to perform a locale-aware query, if duplicate
URLSegments are enabled, to validate the URLSegments for the current
language only.

This is basically a Translatable bug, but it only surfaces in
LanguagePrefix when duplicate entries are enabled. It is to be
considered a temporary fix.